+++ examples/sample_lifecycle.rs
@@ -27,8 +27,8 @@ fn startup(server: Res<AssetServer>, mut commands: Commands) {
 #[derive(Component)]
 struct OnFinished;
 
-// When the sample above completes, `OnRemove` will be triggered on all its
-// components when it gets despawned.
+/// When the sample above completes, [`On<Remove>`] will be triggered on all its
+/// components when it gets despawned.
 fn on_finished(_: On<Remove, OnFinished>, server: Res<AssetServer>, mut commands: Commands) {
     info!("One-shot sample finished, playing looped sample!");